+++ spec/frontend-models/base.http-integration-spec.js
@@ -24,6 +24,7 @@ class HttpFrontendModel extends FrontendModelBase {
         find: "frontend-find",
         index: "frontend-index"
       },
+      modelName: "BrowserFrontendModel",
       path: "/frontend-model-system-tests",
       primaryKey: "id"
     }
@@ -76,6 +77,7 @@ class HttpPreloadComment extends FrontendModelBase {
       commands: {
         index: "frontend-index"
       },
+      modelName: "BrowserFrontendModel",
       path: "/frontend-model-system-tests",
       primaryKey: "id"
     }
@@ -96,6 +98,7 @@ class HttpPreloadTask extends FrontendModelBase {
       commands: {
         index: "frontend-index"
       },
+      modelName: "BrowserFrontendModel",
       path: "/frontend-model-system-tests",
       primaryKey: "id"
     }
@@ -137,6 +140,7 @@ class HttpPreloadProject extends FrontendModelBase {
       commands: {
         index: "frontend-index"
       },
+      modelName: "BrowserFrontendModel",
       path: "/frontend-model-system-tests",
       primaryKey: "id"
     }
@@ -186,7 +190,6 @@ function configureNodeTransport() {
 function configureWebsocketSharedTransport(websocketClient) {
   FrontendModelBase.configureTransport({
     shared: true,
-    url: "/velocious/api",
     websocketClient
   })
 }